+++ apps/frontend/src/auth/auth.config.ts
@@ -4,7 +4,8 @@ import * as Auth from '@dxp/integrations.mocked/auth';
 import { Session, User } from 'next-auth';
 import { AdapterUser } from 'next-auth/adapters';
 import { JWT } from 'next-auth/jwt';
-import { useSession } from 'next-auth/react';
+
+import { sdk } from '@/api/sdk';
 
 const DEFAULT_ROLE = process.env.AUTH_DEFAULT_USER_ROLE!;
 
@@ -37,10 +38,6 @@ export const Adapter = Auth.Adapter;
 export const Providers = Auth.Providers;
 
 export const onSignOut = Auth.signOut;
-
-export async function updateOrganization(session: ReturnType<typeof useSession>) {
-    return Auth.updateOrganization(session);
-}
 
 export const jwtCallback = async (params: JwtCallbackParams): Promise<JWT | null> => {
     return Auth.jwtCallback(() => Promise.resolve(), params, DEFAULT_ROLE);