+++ packages/blocks/document-list/src/api-harmonization/document-list.client.ts
@@ -1,4 +1,4 @@
-export const URL = '/blocks/document-list';
+export { URL } from './document-list.url';
 
 export * as Model from './document-list.model';
 export * as Request from './document-list.request';