+++ packages/blocks/feature-section/src/api-harmonization/feature-section.client.ts
@@ -1,4 +1,4 @@
-export const URL = '/blocks/feature-section';
+export { URL } from './feature-section.url';
 
 export * as Model from './feature-section.model';
 export * as Request from './feature-section.request';