+++ packages/blocks/media-section/src/frontend/MediaSection.types.ts
@@ -7,6 +7,7 @@ export interface MediaSectionProps {
     accessToken?: string;
     locale: string;
     routing: ReturnType<typeof defineRouting>;
+    hasPriority?: boolean;
 }
 
 export type MediaSectionPureProps = MediaSectionProps & Model.MediaSectionBlock;