+++ packages/blocks/pricing-section/src/frontend/PricingSection.server.tsx
@@ -9,7 +9,13 @@ export const PricingSectionDynamic = dynamic(() =>
     import('./PricingSection.client').then((module) => module.PricingSectionPure),
 );
 
-export const PricingSection: React.FC<PricingSectionProps> = async ({ id, accessToken, locale, routing }) => {
+export const PricingSection: React.FC<PricingSectionProps> = async ({
+    id,
+    accessToken,
+    locale,
+    routing,
+    hasPriority,
+}) => {
     try {
         const data = await sdk.blocks.getPricingSection(
             {
@@ -19,7 +25,16 @@ export const PricingSection: React.FC<PricingSectionProps> = async ({ id, access
             accessToken,
         );
 
-        return <PricingSectionDynamic {...data} id={id} accessToken={accessToken} locale={locale} routing={routing} />;
+        return (
+            <PricingSectionDynamic
+                {...data}
+                id={id}
+                accessToken={accessToken}
+                locale={locale}
+                routing={routing}
+                hasPriority={hasPriority}
+            />
+        );
     } catch (error) {
         console.error('Error fetching PricingSection block', error);
         return null;