+++ packages/ui/src/components/Cards/PricingCard/PricingCard.types.ts
@@ -4,6 +4,7 @@ import { Models as FrontendModels } from '@dxp/utils.frontend';
 export type PricingCardProps = {
     image?: Models.Media.Media;
     title: string;
+    titleType?: 'h2' | 'h3' | 'h4';
     description?: string;
     price?: Models.Price.Price;
     tags?: Models.Badge.Badge[];
@@ -13,6 +14,7 @@ export type PricingCardProps = {
     featureListTitle?: string;
     featureList?: Models.FeatureItem.FeatureItem[];
     LinkComponent: FrontendModels.Link.LinkComponent;
+    hasPriority?: boolean;
 };
 
 export type FeatureItemProps = Models.FeatureItem.FeatureItem;